requirements.txt
annotated-types==0.6.0
anyio==4.3.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Js2Py==0.74
openai==1.28.0
packaging==24.0
pipwin==0.5.2
PyAudio==0.2.14
pydantic==2.7.1
pydantic_core==2.18.2
pygame==2.5.2
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dotenv==1.0.1
pyttsx3==2.90
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.10.4
tqdm==4.66.4
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
